embed code in python classes for deleting unused models
