+++ ersilia/cli/commands/__init__.py
@@ -4,17 +4,6 @@ from ... import __version__
 from ... import logger
 from ..echo import Silencer
 
-import json
-import csv
-import time
-
-# in days :
-model_usage_lim   = 30 
-model_cleanup_lim = 7
-
-def seconds_to_days(s):
-    return s / (24 * 3600)
-
 @click.group(cls=BentoMLCommandGroup)
 @click.version_option(version=__version__)
 @click.option(
@@ -43,21 +32,3 @@ def ersilia_cli(verbose, silent):
     silencer.speak()  # To reset default
     if silent:
         silencer.silence()
-    
-    ts_dict = {}
-    if not exists('last_cleaned.json'):
-        ts_dict['timestamp'] = str(time.time())
-        with open('last_cleaned.json', 'w') as outfile:
-            json.dump(ts_dict, outfile)
-    else:
-        current_ts = time.time()
-        with open('last_cleaned.json') as json_file:
-            ts_dict = json.load(json_file)
-            ts = float(ts_dict['timestamp'])
-            if (seconds_to_days(current_ts - ts))>model_cleanup_lim: 
-                with open("fetched_models.txt") as infile:
-                    fetched_models = dict(csv.reader(infile))
-                for m_name in fetched_models: 
-                    if (seconds_to_days(current_ts-float(fetched_models[m_name])))>model_usage_lim:
-                        del_cmd = 'ersilia delete ' + m_name
-                        test_op = os.system(del_cmd)
