+++ ersilia/cli/commands/fetch.py
@@ -42,14 +42,3 @@ def fetch_cmd():
         mf = ModelFetcher(mode=mode, dockerize=dockerize)
         _fetch(mf, model_id)
         echo(":thumbs_up: Model {0} fetched successfully!".format(model_id), fg="green")
-
-        # current timestamp 
-        ts = time.time()
-
-        # write model_name and timestamp to file whenever a new model gets fetched
-        with open("fetched_models.txt","a") as file:
-            file.write(model)
-            file.write(',')
-            file.write(str(ts))
-            file.write('\n')
-
